dockerfiles/python-postgres/requirements.txt
Flask==1.1.2
psycopg2-binary==2.8.6
Jinja2==2.11.3
MarkupSafe==1.1.1
itsdangerous==1.1.0
werkzeug==1.0.1

